Add a mechanism for allowing code to check if it's in a unit test.
This is protected from general use/abuse through friendship. Generally
a bad idea to use this mechanism, and it has explicitly been made hard
to use through requiring friendship and a long class name to type that
makes you state you're aware that it's a bad idea before you can use
it.