jhm: Try can fail / return nullptr
This particular case gets hit when we're trying to cache finish a file
which no longer exists and can't be produced. The cache files refer
to it, which causes us to try to cache complete it. We should run into
the fact it doesn't exist and can't be produced. Calling IsDone() with
a nullptr though is a bad idea.