base/cmd: When printing missing required arguments, only print the same argument once.
The ByName vector of parameters which wa being iterated over contained the same parameter multiple times if it had multiple forms it could be given in (-h, --help, -?). Now we unique the parameters when checking if all the required parameters have been given, so that we don't print the required parameter once per alternate method of giving it.
 the parameter

Fixes #236